Moved DragAndDropFileInput to separate folder
